requirements.txt
openai
whisper
srt
yt-dlp
python-dotenv
ffmpeg-python
streamlit